Moved SampleClass, ISampleRepository and IMinimalRepository to a new SamplaLibrary library project
