+++ Samples/SampleLibrary/Services/IMinimalRepository.cs
@@ -1,0 +1,9 @@
+﻿using QTFK.Services;
+using SampleLibrary.Models;
+
+namespace SampleLibrary.Services
+{
+    public interface IMinimalRepository : IRepository<SampleClass>
+    {
+    }
+}
