+++ Tests/QTFK.Data.Tests/QTFK.Data.Tests.csproj
@@ -74,9 +74,6 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="Models\IMinimalRepository.cs" />
-    <Compile Include="Models\ISampleRepository.cs" />
-    <Compile Include="Models\SampleClass.cs" />
     <Compile Include="Models\SampleRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="NewTest.cs" />
@@ -90,6 +87,10 @@
     <ProjectReference Include="..\..\QTFK.Data\QTFK.Data.csproj">
       <Project>{a78ecf12-d6e8-4c49-b315-ffb4b3fb11cc}</Project>
       <Name>QTFK.Data</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\Samples\SampleLibrary\SampleLibrary.csproj">
+      <Project>{ca2aa41a-1717-47ca-b528-b67a646b1afa}</Project>
+      <Name>SampleLibrary</Name>
     </ProjectReference>
   </ItemGroup>
   <Choose>
